+++ epam-promo/components/inputs/MultiSwitch.tsx
@@ -7,6 +7,12 @@ type MultiSwitchMods = {
      * @default 'blue'
      */
     color?: 'blue' | 'gray50' | 'gray' | uui.MultiSwitchProps['color'];
+
+    /**
+     * Defines component size.
+     * @default '36'
+     */
+    size?: '24' | '30' | '36' | '42' | '48' | '60';
 };
 
 /** Represents the properties for the MultiSwitch component. */